+++ CHANGELOG.md
@@ -5,6 +5,7 @@ https://keepachangelog.com/en/1.0.0/
 - `StringConcatenatedInLoop`: No longer triggers for regular assignments
 - `UnnecessaryEnumerableMaterialization`: now supports conditional access, i.e. `values?.ToArray().ToList()`
 - `AsyncOverloadsAvailable`: No longer triggers when contained within a `lock` body
+- `AsyncOverloadsAvailable`: Better maintains whitespace and comments when the code fix is applied
 - `ExplicitEnumValues` and `EnumWithoutDefaultValue`: Downgraded from warning to info to reduce their prevalence
 - `ExceptionThrownFromProhibitedContext`: No longer triggers for `PlatformNotSupportedException`
 - `StringPlaceholdersInWrongOrder` and `UnnecessaryEnumerableMaterialization`: Rewritten to use `IOperation`